+++ firmware/src/plant/fixture/plant_structure.h
@@ -22,14 +22,14 @@
 
 /* The one coefficient this structure reads. It means nothing. */
 typedef struct {
-    double fixture_gain;
+    float fixture_gain;
 } plant_parameters_t;
 
 /* One accumulator, and the record it was initialised from. */
 typedef struct {
     bool initialised;
     plant_parameters_t coefficients;
-    double accumulated;
+    float accumulated;
 } plant_model_t;
 
 /* The single coefficient this structure requires, and its admissible range. */
@@ -39,6 +39,6 @@ const plant_parameter_spec_t *plant_structure_parameter_specs(size_t *count);
  * Add what the brew heater commanded over `seconds`, scaled by the gain. This
  * is the whole of the structure's behaviour.
  */
-void fixture_accumulate(plant_model_t *model, const plant_actuation_t *actuation, double seconds);
+void fixture_accumulate(plant_model_t *model, const plant_actuation_t *actuation, float seconds);
 
 #endif /* PLANT_STRUCTURE_H */
